# docs/api/v1/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Xenoteer Control API",
    "version": "1.0.0-phase4",
    "description": "Implemented Phase 4 HTTP control, observation, clipboard, capture, artifact, and viewer contract. Typed request and result definitions are generated from xenoteer-protocol and referenced from schemas/v1. Control WebSocket application messages are documented separately in websocket.md.",
    "license": {
      "name": "Apache License 2.0",
      "identifier": "Apache-2.0"
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:3000",
      "description": "Default container-local control plane; production TLS may terminate at a trusted same-host proxy."
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/v1/status": {
      "get": {
        "operationId": "getStatus",
        "tags": ["Discovery"],
        "summary": "Read detailed server, desktop, and capability status",
        "responses": {
          "200": {
            "description": "Current authenticated status snapshot.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                },
                "examples": {
                  "ready": {
                    "externalValue": "./examples/status-response.json"
                  }
                }
              }
            }
          },
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "429": {"$ref": "#/components/responses/TooManyRequests"},
          "504": {"$ref": "#/components/responses/RequestOutcomeUnknown"}
        }
      }
    },
    "/v1/capabilities": {
      "get": {
        "operationId": "getCapabilities",
        "tags": ["Discovery"],
        "summary": "Read the deterministic capability report",
        "responses": {
          "200": {
            "description": "Current capability report sorted by identifier.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../../../schemas/v1/capabilities.json"
                },
                "examples": {
                  "ready": {
                    "externalValue": "./examples/capabilities-response.json"
                  }
                }
              }
            }
          },
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "429": {"$ref": "#/components/responses/TooManyRequests"},
          "504": {"$ref": "#/components/responses/RequestOutcomeUnknown"}
        }
      }
    },
...
    "/v1/desktops/{desktop_id}/commands": {
      "parameters": [
        {"$ref": "#/components/parameters/DesktopId"}
      ],
      "post": {
        "operationId": "submitCommand",
        "tags": ["Commands"],
        "summary": "Submit or idempotently retrieve a command",
        "description": "The optional Idempotency-Key must equal command_id. A disconnect or transport timeout never cancels accepted work.",
        "parameters": [
          {"$ref": "#/components/parameters/IdempotencyKey"}
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "../../../schemas/v1/command-envelope.json"},
              "examples": {
                "smoothPointerMove": {"externalValue": "./examples/command-submit-request.json"}
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "New or existing command is nonterminal.",
            "headers": {
              "Location": {"$ref": "#/components/headers/Location"},
              "Retry-After": {"$ref": "#/components/headers/RetryAfter"}
            },
            "content": {
              "application/json": {
                "schema": {"$ref": "../../../schemas/v1/command-result.json"},
                "examples": {
                  "accepted": {"externalValue": "./examples/command-accepted-response.json"}
                }
              }
            }
          },
          "200": {
            "description": "Exact command ID and content were already terminal.",
            "headers": {
              "Location": {"$ref": "#/components/headers/Location"}
            },
            "content": {
              "application/json": {
                "schema": {"$ref": "../../../schemas/v1/command-result.json"},
                "examples": {
                  "succeeded": {"externalValue": "./examples/command-succeeded-response.json"}
                }
              }
            }
          },
          "400": {"$ref": "#/components/responses/BadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "404": {"$ref": "#/components/responses/NotFound"},
          "409": {"$ref": "#/components/responses/Conflict"},
          "413": {"$ref": "#/components/responses/PayloadTooLarge"},
          "422": {"$ref": "#/components/responses/UnsupportedByTarget"},
          "429": {"$ref": "#/components/responses/TooManyRequests"},
          "503": {"$ref": "#/components/responses/CapabilityUnavailable"},
          "504": {"$ref": "#/components/responses/RequestOutcomeUnknown"}
        }
      }
    },
...
    "/v1/ws": {
      "get": {
        "operationId": "upgradeWebSocket",
        "tags": ["WebSocket"],
        "summary": "Upgrade to the v1 text-JSON application protocol",
        "description": "The HTTP exchange is an RFC 6455 upgrade. See websocket.md for implemented message shapes, including owner-scoped process.exited events and global metadata-free resynchronization barriers. Browser Origin is checked against an exact allowlist; non-browser clients may omit it.",
        "responses": {
          "101": {
            "description": "Switching Protocols."
          },
          "400": {"$ref": "#/components/responses/BadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "429": {"$ref": "#/components/responses/TooManyRequests"}
        },
        "x-xenoteer-websocket-document": "./websocket.md"
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Exactly one long-lived API Bearer credential. The token is opaque and must not be placed in URLs. The closed grant vocabulary is desktop:status, desktop:observe, input:control, application:launch, application:terminate, window:control, clipboard:read, clipboard:write, capture:read, artifact:read, artifact:delete, and viewer:read. This scheme is not accepted by the viewer gateway WebSocket."
      },
      "viewerTicketSubprotocol": {
        "type": "apiKey",
        "in": "header",
        "name": "Sec-WebSocket-Protocol",
        "description": "A short-lived single-use viewer ticket carried only as the xenoteer.ticket.<secret> WebSocket subprotocol alongside the exact binary subprotocol. Never place the secret in a URL or log it."
      }
    },
    "parameters": {
      "DesktopId": {
        "name": "desktop_id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
...
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": false,
        "description": "When present exactly once, must equal the textual command_id from the body.",
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      }
    },
    "headers": {
      "Location": {
        "description": "Canonical command, caller-owned lease, or private artifact resource.",
        "schema": {"type": "string"}
      },
      "RetryAfter": {
        "description": "Current implementation emits 1 second for nonterminal command results and backoff responses.",
        "required": true,
        "schema": {"type": "integer", "minimum": 1, "example": 1}
      },
      "OptionalRetryAfter": {
        "description": "Present with value 1 while the returned command result is nonterminal; absent when terminal.",
        "schema": {"type": "integer", "minimum": 1, "example": 1}
      }
    },
    "schemas": {
...
      "ProtocolVersion": {
        "type": "object",
        "additionalProperties": false,
        "required": ["major", "minor"],
        "properties": {
          "major": {"type": "integer", "minimum": 0, "maximum": 65535},
          "minor": {"type": "integer", "minimum": 0, "maximum": 65535}
        }
      },
      "DesktopStatus": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "generation", "state", "reason_code"],
        "properties": {
          "id": {"type": "string", "format": "uuid"},
          "generation": {"type": ["string", "null"], "format": "uuid"},
          "state": {
            "type": "string",
            "enum": ["booting", "probing", "ready", "degraded", "draining", "stopped", "failed"]
          },
          "reason_code": {"type": ["string", "null"]}
        }
      },
      "StatusResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["server_version", "protocol_min", "protocol_max", "desktop", "capabilities"],
        "properties": {
          "server_version": {"type": "string"},
          "protocol_min": {"$ref": "#/components/schemas/ProtocolVersion"},
          "protocol_max": {"$ref": "#/components/schemas/ProtocolVersion"},
          "desktop": {"$ref": "#/components/schemas/DesktopStatus"},
          "capabilities": {"$ref": "../../../schemas/v1/capabilities.json"}
        }
      }
    },
    "responses": {
...
      "BadRequest": {
        "description": "Malformed JSON, invalid path/query/header, strict-shape failure, or unsupported protocol major.",
        "content": {
          "application/problem+json": {
            "schema": {"$ref": "../../../schemas/v1/problem.json"}
          }
        }
      },
      "Unauthorized": {
        "description": "A valid Bearer credential is required.",
        "headers": {
          "WWW-Authenticate": {
            "schema": {"type": "string", "const": "Bearer realm=\"xenoteer\""}
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {"$ref": "../../../schemas/v1/problem.json"}
          }
        }
      },
      "Forbidden": {
        "description": "The principal lacks the required grant, or a browser WebSocket Origin is denied.",
        "content": {
          "application/problem+json": {
            "schema": {"$ref": "../../../schemas/v1/problem.json"}
          }
        }
      },
      "NotFound": {
        "description": "The requested versioned resource does not exist or may not be disclosed.",
        "content": {
          "application/problem+json": {
            "schema": {"$ref": "../../../schemas/v1/problem.json"}
          }
        }
      },
      "Conflict": {
        "description": "Stale generation, lease conflict, command ID/content conflict, or unsafe cancellation stage.",
        "headers": {
          "Retry-After": {"$ref": "#/components/headers/OptionalRetryAfter"}
        },
        "content": {
          "application/problem+json": {
            "schema": {"$ref": "../../../schemas/v1/problem.json"},
            "examples": {
              "staleReference": {"externalValue": "./examples/problem-stale-reference.json"}
            }
          }
        }
      },
      "PayloadTooLarge": {
        "description": "The request body exceeds the configured ordinary-message byte limit.",
        "content": {
          "application/problem+json": {
            "schema": {"$ref": "../../../schemas/v1/problem.json"}
          }
        }
      },
      "UnsupportedByTarget": {
        "description": "The target cannot perform the structurally valid operation.",
        "content": {
          "application/problem+json": {
            "schema": {"$ref": "../../../schemas/v1/problem.json"}
          }
        }
      },
      "TooManyRequests": {
        "description": "A bounded admission slot, queue, rate limit, or quota is full.",
        "headers": {
          "Retry-After": {"$ref": "#/components/headers/RetryAfter"}
        },
        "content": {
          "application/problem+json": {
            "schema": {"$ref": "../../../schemas/v1/problem.json"}
          }
        }
      },
      "CapabilityUnavailable": {
        "description": "The desktop or required subsystem is not ready.",
        "headers": {
          "Retry-After": {"$ref": "#/components/headers/RetryAfter"}
        },
        "content": {
          "application/problem+json": {
            "schema": {"$ref": "../../../schemas/v1/problem.json"}
          }
        }
      },
      "RequestOutcomeUnknown": {
        "description": "The handler deadline elapsed after dispatch. This does not cancel accepted work.",
        "content": {
          "application/problem+json": {
            "schema": {"$ref": "../../../schemas/v1/problem.json"},
            "examples": {
              "exactCommandRetry": {"externalValue": "./examples/problem-request-outcome-unknown.json"},
              "noAutomaticRetry": {"externalValue": "./examples/problem-request-outcome-unknown-no-retry.json"}
            }
          }
        }
      }
    }
  }
}